+++ tests/test_dictfield.py
@@ -15,87 +15,45 @@ from . import test_settings
 
 from datetime import date
 
-from django.core.exceptions import ValidationError
+from rest_framework.serializers import ValidationError
 from rest_framework import ISO_8601
-from rest_framework.fields import CharField
-from rest_framework.fields import DateField
+from rest_framework.serializers import CharField
+from rest_framework.serializers import DateField
 import pytest
 
 from drf_compound_fields.fields import DictField
 
 
-def test_from_native_no_value_field():
+def test_to_internal_value_with_child():
     """
-    When a DictField has no value-field, from_native should return the data it was given
-    un-processed.
-    """
-    field = DictField()
-    data = {"a": 1, "b": 2}
-    obj = field.from_native(data)
-    assert data == obj
-
-
-def test_to_native_no_value_field():
-    """
-    When a DictField has no value-field, to_native should return the data it was given
-    un-processed.
-    """
-    field = DictField()
-    obj = {"a": 1, "b": 2}
-    data = field.to_native(obj)
-    assert obj == data
-
-
-def test_from_native_with_value_field():
-    """
-    When a DictField has an value-field, from_native should return a dict of elements resulting
-    from the application of the value-field's from_native method to each value of the input
+    When a DictField has an value-field, to_internal_value should return a dict of elements resulting
+    from the application of the value-field's to_internal_value method to each value of the input
     data dict.
     """
-    field = DictField(DateField())
+    field = DictField(child=DateField())
     data = {"a": "2000-01-01", "b": "2000-01-02"}
-    obj = field.from_native(data)
+    obj = field.to_internal_value(data)
     assert {"a": date(2000, 1, 1), "b": date(2000, 1, 2)} == obj
 
 
-def test_to_native_with_value_field():
+def test_to_representation_with_child():
     """
-    When a DictField has an value-field, to_native should return a dict of elements resulting from
-    the application of the value-field's to_native method to each value of the input object dict.
+    When a DictField has an value-field, to_representation should return a dict of elements resulting from
+    the application of the value-field's to_representation method to each value of the input object dict.
     """
-    field = DictField(DateField(format=ISO_8601))
+    field = DictField(child=DateField(format=ISO_8601))
     obj = {"a": date(2000, 1, 1), "b": date(2000, 1, 2)}
-    data = field.to_native(obj)
+    data = field.to_representation(obj)
     assert {"a": "2000-01-01", "b": "2000-01-02"} == data
-
-
-def test_missing_required_dict():
-    """
-    When a DictField requires a value, then validate should raise a ValidationError on a missing
-    (None) value.
-    """
-    field = DictField()
-    with pytest.raises(ValidationError):
-        field.validate(None)
 
 
 def test_validate_non_dict():
     """
     When a DictField is given a non-dict value, then validate should raise a ValidationError.
     """
-    field = DictField()
+    field = DictField(child=DateField())
     with pytest.raises(ValidationError):
-        field.validate('notADict')
-
-
-def test_validate_empty_dict():
-    """
-    When a DictField requires a value, then validate should raise a ValidationError on an empty
-    value.
-    """
-    field = DictField()
-    with pytest.raises(ValidationError):
-        field.validate({})
+        field.to_internal_value('notADict')
 
 
 def test_validate_elements_valid():
@@ -103,9 +61,9 @@ def test_validate_elements_valid():
     When a DictField is given a dict whose values are valid for the value-field, then validate
     should not raise a ValidationError.
     """
-    field = DictField(CharField(max_length=5))
+    field = DictField(child=CharField(max_length=5))
     try:
-        field.run_validators({"a": "a", "b": "b", "c": "c"})
+        field.to_internal_value({"a": "a", "b": "b", "c": "c"})
     except ValidationError:
         assert False, "ValidationError was raised"
 
@@ -115,18 +73,6 @@ def test_validate_elements_invalid():
     When a DictField is given a dict containing values that are invalid for the value-field, then
     validate should raise a ValidationError.
     """
-    field = DictField(CharField(max_length=5))
+    field = DictField(child=CharField(max_length=5))
     with pytest.raises(ValidationError):
-        field.run_validators({"a": "012345", "b": "012345"})
-
-def test_missing_not_required_dict():
-    """
-    When a DictField do not require a value, then validate should not raise a
-    ValidationError on a missing (None) value.
-    """
-    field = DictField(required=False)
-
-    try:
-       field.validate(None)
-    except ValidationError:
-       assert False, "ValidationError was raised"
+        field.to_internal_value({"a": "012345", "b": "012345"})